+++ src/App.jsx
@@ -5,8 +5,8 @@ import Navbar from './components/Navbar';
 import SplashScreen from './components/SplashScreen';
 import MyReservations from './pages/MyReservations';
 import Reserve from './pages/Reserve';
-import AddProperty from './components/AddProperty';
-import DeleteProperty from './components/DeleteProperty';
+import AddProperty from './pages/AddProperty';
+import DeleteProperty from './pages/DeleteProperty';
 import HouseDetail from './pages/HouseDetail';
 import Mainpage from './pages/Mainpage';
 import './App.css';
