--- src/components/Home.jsx
@@ -1,9 +1,0 @@
-import './stylesheets/Login.css';
-
-function Home() {
-  return (
-    <h1>All Properties</h1>
-  );
-}
-
-export default Home;
