+++ src/pages/stylesheets_page/DeleteProperty.css
